feat: prevent items or collection to be deleted

+++ src/Collection/Collection.router.ts
@@ -49,7 +49,7 @@ export class CollectionRouter extends Router {
     this.router.put(
       '/collections/:id',
       withAuthentication,
-      server.handleRequest(this.upsertCollection)
+      server.handleRequest(this.upsertCollection.bind(this))
     )
 
     /**
@@ -60,7 +60,7 @@ export class CollectionRouter extends Router {
       withAuthentication,
       withCollectionExists,
       withCollectionAuthorization,
-      server.handleRequest(this.deleteCollection)
+      server.handleRequest(this.deleteCollection.bind(this))
     )
   }
 
@@ -121,6 +121,16 @@ export class CollectionRouter extends Router {
         )
       }
 
+      if (await this.isCollectionPublished(id)) {
+        throw new HTTPError(
+          "The collection can't be updated because it is published",
+          {
+            id,
+          },
+          STATUS_CODES.unauthorized
+        )
+      }
+
       const attributes = {
         ...collectionJSON,
         eth_address,
@@ -148,10 +158,37 @@ export class CollectionRouter extends Router {
 
   async deleteCollection(req: AuthRequest) {
     const id = server.extractFromReq(req, 'id')
+
+    if (await this.isCollectionPublished(id)) {
+      throw new HTTPError(
+        "The collection was published. It can't be deleted",
+        {
+          id,
+        },
+        STATUS_CODES.unauthorized
+      )
+    }
+
     await Promise.all([
       Collection.delete({ id }),
       Item.delete({ collection_id: id }),
     ])
     return true
+  }
+
+  async isCollectionPublished(collectionId: string) {
+    const dbCollection = await Collection.findOne<CollectionAttributes>({
+      id: collectionId,
+    })
+
+    if (!dbCollection) {
+      return false
+    }
+
+    const remoteCollections = await collectionAPI.fetchCollection(
+      dbCollection.contract_address
+    )
+
+    return !!remoteCollections
   }
 }
